# libplugin-pay: use map for channel hints