@cdecker 2024-10-07
Thanks @JssDWt for rebuilding this on top of #7494, and sorry for the delay.

ACK https://github.com/ElementsProject/lightning/pull/7726/commits/83d1db07ec0b160b03d88a4210f61ecd22333087